+++ assets/Assets.php
@@ -15,7 +15,7 @@ class Assets extends AssetBundle
 {
 
     public $publishOptions = [
-        'forceCopy' => true
+        'forceCopy' => false
     ];
 
     public $sourcePath = '@tasks/resources';